Create StateOfCharge.ino
This is to be run on the photon that collects state of charge data from the batteries, as well as current and voltage data going into the system to keep track of power generated and consumed. The code will need to be expanded for the added funtionality when the current and data sensors are ordered and brough online.